backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
geopy>=2.3.0
requests>=2.31.0
tenacity>=8.2.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
jinja2>=3.1.0
